requirements.txt
requests
fastapi
pydantic==1.10.10
python-jose
requests_oauthlib
gspread
botocore
stripe
uvicorn[standard]
passwordgenerator
python-terraform
airpress
cffi
openstacksdk
asyncio
pyopenssl==22.1.0
black==23.9.1
websockets==11.0.3
